New functions and methods added
Default values for Vector constructor were set 0.0.
operator+ impelented beyond struct Vector; operator-=,
operator*=, operator- and operator* implemented.

Functions calculating scalar, vector and triple products implemented.
Functions that check if vectors are equal, collinear or complanar
implemented. Calculating the length of a vector and checking if it is
equal to the null vector are implemented as member functions.