Class Plane, member and other functions coded General equation (Ax + By + Cz + D = 0) of a plane is its representation
Member-functions: 2 types of constructors: 1) from general form of plane (by A, B, С, D); 2) from origin point and normal vector. Also there are functions of receiving normal vector and origin point.

There are functions of mutual location of plane and (point/line/plane) in public access: 1) For plane and point: distance between them and function of checking beloning a point to a plane; 2) For plane and line: functions of checking if plane and line parallel, intersect or line is consisted in plane. Also function of distance between parallel plane and line and function returing a point of intersectior were implemented; 3) For plane and plane: functions of checking if planes are parallel, equal, intersect. Also for parallel and equal planes was added function of distance beetween them.